// app/src/generated-api/models/CreateServiceOrdersRequest.ts
export interface CreateServiceOrdersRequest {
    /**
     * ID do veículo ao qual a ordem de serviço está associada.
     * @type {string}
     * @memberof CreateServiceOrdersRequest
     */
    vehicleId: string;
    /**
     * ID do usuário que inseriu/criou a ordem de serviço.
     * @type {string}
     * @memberof CreateServiceOrdersRequest
     */
    userId: string;
    /**
     * Descrição detalhada do problema ou serviço a ser realizado.
     * @type {string}
     * @memberof CreateServiceOrdersRequest
     */
    description: string;
    /**
     * 
     * @type {Type}
     * @memberof CreateServiceOrdersRequest
     */
    type: Type;
    /**
     * 
     * @type {Category}
     * @memberof CreateServiceOrdersRequest
     */
    category: Category;
    /**
     * 
     * @type {MaintenanceNature}
     * @memberof CreateServiceOrdersRequest
     */
    maintenanceNature: MaintenanceNature;
    /**
     * 
     * @type {Impact}
     * @memberof CreateServiceOrdersRequest
     */
    impact: Impact;
    /**
     * 
     * @type {Priority}
     * @memberof CreateServiceOrdersRequest
     */
    priority: Priority;
    /**
     * Data em que o serviço foi agendado ou concluído.
     * @type {Date}
     * @memberof CreateServiceOrdersRequest
     */
    serviceDate?: Date | null;
    /**
     * 
     * @type {Status}
     * @memberof CreateServiceOrdersRequest
     */
    status?: Status;
}