// app/src/generated-api/models/Status.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Travel Manager - API de Gestão de recursos de transporte rodoviário
 * Especificação modular da API de back-end. Todos os schemas e paths estão em ficheiros separados (specs/). Use um agregador OpenAPI (como o ReDoc ou Swagger UI) para visualizar a documentação completa.
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: ferdxavier@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Status atual da ordem de serviço.
 * @export
 */
export const Status = {
    Open: 'open',
    InProgress: 'in_progress',
    Closed: 'closed',
    Canceled: 'canceled'
} as const;
export type Status = typeof Status[keyof typeof Status];


export function instanceOfStatus(value: any): boolean {
    for (const key in Status) {
        if (Object.prototype.hasOwnProperty.call(Status, key)) {
            if (Status[key as keyof typeof Status] === value) {
                return true;
            }
        }
    }
    return false;
}

export function StatusFromJSON(json: any): Status {
    return StatusFromJSONTyped(json, false);
}

export function StatusFromJSONTyped(json: any, ignoreDiscriminator: boolean): Status {
    return json as Status;
}

export function StatusToJSON(value?: Status | null): any {
    return value as any;
}

export function StatusToJSONTyped(value: any, ignoreDiscriminator: boolean): Status {
    return value as Status;
}